[#69990] 댓글 & 답글 작성 model 구현
- 댓글의 갯수를 알아오는 쿼리 추가
- 댓글 추가 쿼리 추가
- 답글 추가시 업데이트 해주는 쿼리 추가
- 변수명들 int-> Long으로 변경

+++ src/com/zum/pilot/dao/PostDao.java
@@ -18,8 +18,8 @@ public class PostDao {
 	}
 	
 	// 게시글의 총 갯수
-	public int totalNumberOfPost() {
-		int totalCount = 0;
+	public Long totalNumberOfPost() {
+		Long totalCount = 0L;
 		Connection con = null;
 		ResultSet rs = null;
 		PreparedStatement pstmt = null;
@@ -32,7 +32,7 @@ public class PostDao {
 			rs = pstmt.executeQuery();
 
 			if (rs.next()) {
-				totalCount = rs.getInt(1);
+				totalCount = rs.getLong(1);
 			}
 
 		} catch (SQLException e) {
@@ -54,6 +54,7 @@ public class PostDao {
 		}
 		return totalCount;
 	}
+	
 
 	// 게시글 불러오기
 	public List<PostVo> getList(int currentPageNum, int postUnit) {
